src/scrapers/linkedin_scraper.py
"""
LinkedIn Company Page Discovery Module
Uses ethical scraping methods to discover new LinkedIn company pages
"""

import requests
import time
import random
from typing import List, Dict, Any, Optional
from bs4 import BeautifulSoup
from datetime import datetime
from src.core.config_manager import config
from src.core.database import DatabaseHandler
from src.core.lead_scorer import LeadScorer
from src.core.logger import get_logger

logger = get_logger(__name__)


class LinkedInScraper:
    """Discover newly created LinkedIn Company Pages for Nigerian businesses"""
    
    def __init__(self, db_handler: DatabaseHandler):
        self.db = db_handler
        self.scorer = LeadScorer()
        
        # Load configuration
        self.user_agents = config.get_user_agents()
        self.rate_limit = config.get_rate_limit('linkedin')
        
        # Session for requests
        self.session = requests.Session()
        
        logger.info("LinkedIn scraper initialized")
    
    def _get_headers(self) -> Dict[str, str]:
        """Get randomized headers for requests"""
        return {
            'User-Agent': random.choice(self.user_agents),
            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
            'Accept-Language': 'en-US,en;q=0.5',
            'Accept-Encoding': 'gzip, deflate, br',
            'DNT': '1',
            'Connection': 'keep-alive',
            'Upgrade-Insecure-Requests': '1'
        }
    
    def search_companies(self, query: str, location: str = "Nigeria") -> List[Dict[str, Any]]:
        """
        Search for LinkedIn company pages
        
        Note: LinkedIn heavily restricts scraping. Ethical approaches:
        1. Use LinkedIn's official API (requires partnership)
        2. Use Google search with site:linkedin.com/company
        3. Use public company directory pages
        4. Manual monitoring with automation assistance
        
        This implementation uses method #2 (Google search)
        """
        results = []
        
        try:
            logger.info(f"Searching LinkedIn for: {query} in {location}")
            
            # We'll use Google search to find LinkedIn company pages
            # This is more reliable than direct LinkedIn scraping
            search_query = f'site:linkedin.com/company {query} {location}'
            
            # This will be handled by the Google scraper
            # For now, we'll return empty and let Google scraper handle it
            
            self._rate_limit_delay()
            
        except Exception as e:
            logger.error(f"Error searching LinkedIn: {str(e)}")
        
        return results
    
    def get_company_details(self, company_url: str) -> Optional[Dict[str, Any]]:
        """
        Get details from a LinkedIn company page
        
        Note: This uses public page scraping. LinkedIn may block requests.
        Use with caution and respect rate limits.
        """
        try:
            headers = self._get_headers()
            response = self.session.get(company_url, headers=headers, timeout=30)
            
            if response.status_code == 200:
                soup = BeautifulSoup(response.content, 'html.parser')
                
                # Extract company information
                business_data = self._extract_company_info(soup, company_url)
                
                return business_data
            else:
                logger.warning(f"Failed to fetch {company_url}: Status {response.status_code}")
                return None
                
        except Exception as e:
            logger.error(f"Error getting company details from {company_url}: {str(e)}")
            return None
    
    def _extract_company_info(self, soup: BeautifulSoup, url: str) -> Dict[str, Any]:
        """
        Extract company information from LinkedIn page HTML
        
        Note: LinkedIn's HTML structure changes frequently.
        This is a simplified version that may need updates.
        """
        business_data = {
            'business_name': '',
            'platform': 'LinkedIn',
            'page_url': url,
            'category': '',
            'location': '',
            'phone': '',
            'email': '',
            'description': '',
            'page_created_date': '',
            'metadata': {}
        }
        
        try:
            # Company name
            name_elem = soup.find('h1', class_='org-top-card-summary__title')
            if name_elem:
                business_data['business_name'] = name_elem.get_text(strip=True)
            
            # Description
            desc_elem = soup.find('p', class_='org-top-card-summary__tagline')
            if desc_elem:
                business_data['description'] = desc_elem.get_text(strip=True)
            
            # Location
            loc_elem = soup.find('div', class_='org-top-card-summary-info-list__info-item')
            if loc_elem:
                business_data['location'] = loc_elem.get_text(strip=True)
            
            # Industry (category)
            industry_elem = soup.find('dd', class_='org-page-details__definition-text')
            if industry_elem:
                business_data['category'] = industry_elem.get_text(strip=True)
            
            # Company size
            size_elem = soup.find_all('dd', class_='org-page-details__definition-text')
            if len(size_elem) > 1:
                business_data['metadata']['company_size'] = size_elem[1].get_text(strip=True)
            
        except Exception as e:
            logger.error(f"Error extracting company info: {str(e)}")
        
        return business_data
    
    def discover_new_companies(self) -> List[Dict[str, Any]]:
        """
        Main discovery method - searches for new company pages
        
        Note: Due to LinkedIn's restrictions, this method primarily
        relies on Google search integration (handled by google_scraper.py)
        """
        discovered = []
        
        logger.info("LinkedIn discovery relies on Google search integration")
        logger.info("Use the Google scraper with LinkedIn-specific queries")
        
        # In practice, you would:
        # 1. Get company URLs from Google search results
        # 2. Fetch details for each company
        # 3. Score and store new companies
        
        return discovered
    
    def process_company_url(self, url: str, category: str = None) -> Optional[Dict[str, Any]]:
        """
        Process a single LinkedIn company URL
        This is called by the Google scraper when it finds LinkedIn pages
        """
        # Check if already in database
        if self.db.business_exists(url):
            logger.debug(f"Company already in database: {url}")
            return None
        
        # Get company details
        business_data = self.get_company_details(url)
        
        if not business_data or not business_data.get('business_name'):
            return None
        
        # Set category if provided
        if category:
            business_data['category'] = category
        
        # Score the lead
        scored_data = self.scorer.score_business(business_data)
        
        # Check minimum confidence
        min_score = config.get_detection_settings().get('min_confidence_score', 60)
        if scored_data['confidence_score'] >= min_score:
            # Add to database
            if self.db.add_business(scored_data):
                logger.info(f"New LinkedIn company discovered: {scored_data['business_name']}")
                return scored_data
        
        return None
    
    def _rate_limit_delay(self):
        """Add delay between requests to respect rate limits"""
        delay = random.uniform(
            self.rate_limit['delay_min'],
            self.rate_limit['delay_max']
        )
        time.sleep(delay)

